WebIDL/tests/submissions/heycam/long-003.html

<!DOCTYPE html>
<title>long</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.8 long
//
// IDL long values are represented as integer JS Number values in the
// range [-2147483648, 2147483647].

// feature: HTMLPreElement.width (attribute) (of type long)

test(function() {
  var e = document.createElement("pre");
  assert_equals(typeof e.width, "number");
  assert_true(e.width == Math.floor(e.width) && e.width >= -2147483648 && e.width <= 2147483647);
}, "HTMLPreElement.width returns a value of the correct type");

// feature: HTMLElement.tabIndex (attribute) (of type long)

test(function() {
  var e = document.createElement("span");
  assert_equals(typeof e.tabIndex, "number");
  assert_true(e.tabIndex == Math.floor(e.tabIndex) && e.tabIndex >= -2147483648 && e.tabIndex <= 2147483647);
}, "HTMLElement.tabIndex returns a value of the correct type");
</script>
